Move types definition file to /app directory, remove global directory
Move types definition file to /app directory, remove global directory
