+++ src/components/home/components/forms/AddMappingForm.tsx
@@ -4,7 +4,7 @@ import cloneDeep from 'lodash/cloneDeep';
 import { Row, Col } from 'react-bootstrap';
 
 /* Import Types */
-import { Dict } from 'global/Types';
+import { Dict } from 'app/Types';
 
 /* Import Styling */
 import styles from 'components/home/home.module.scss';
