+++ src/components/home/components/overview/MASOverview.tsx
@@ -7,7 +7,7 @@ import { getMachineAnnotationServices } from 'redux/MAS/MASSlice';
 import { setEditTarget } from 'redux/edit/EditSlice';
 
 /* Import Types */
-import { EditTarget, Dict } from 'global/Types';
+import { EditTarget, Dict } from 'app/Types';
 
 /* Import Components */
 import OverviewTable from './table/OverviewTable';
