+++ src/components/home/components/overview/MappingsOverview.tsx
@@ -7,7 +7,7 @@ import { getMappings } from 'redux/mapping/MappingSlice';
 import { setEditTarget } from 'redux/edit/EditSlice';
 
 /* Import Types */
-import { EditTarget, Dict } from 'global/Types';
+import { EditTarget, Dict } from 'app/Types';
 
 /* Import Components */
 import OverviewTable from './table/OverviewTable';
